CustomLinkedList

+++ src/manager/HistoryManager.java
@@ -8,7 +8,7 @@ public interface HistoryManager {
 
     void add(Task task);
 
-    void remove(int id);
+    void remove(long id);
 
     List<Task> getHistory();
